Initial checkin of Verilog Analyzer 2 binary.
Starting set of files for the va2 binary. Given an initial set of Verilog module names, this binary
will perform a recursive search for Verilog source files. The existing Verible APIs require either
all Verilog files in a single directory or all Verilog files listed in a manifest file. This binary
use the -v/-y/-I convention to search for existing Verilog files.

Still to be done:
- Add Verilog library processing
- Add other special files
- Add YAML output capability for later information processing

NOTE: The skeleton files here are slightly overdone, but make for a decent baseline of how to build
a command line tool using the Verible API.

NOTE 2025/01/27: Updated for new Verible file locations.

[skip ci]